Re-did the dataset to be based on max number of song events
Previously we cut songs based on a maximum length in seconds, not we cap them based on maximum number of song events which should keep the data length the same regardless of the number of instruments used. consolidated_data_600_events.jsonl is not uploaded because its above 100mb. You can generate it yourself through ConsolidateData.py